[*] update `main()` to support `serve` command to starts a dns server
- move `main.go` under `cmd/`
